release version 2.1.3

+++ Models/BacklinksIdListResultInfo.cs
@@ -52,7 +52,7 @@ namespace DataForSeo.Client.Models
         /// total tasks cost, USD
         /// </summary>
         [JsonProperty("cost", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? Cost { get; set; }
+        public int? Cost { get; set; }
 
         /// <summary>
         /// contains parameters you specified in the POST request
